requirements.txt
tortoise-orm==0.15.5
aiosqlite==0.10.0
asyncpg==0.20.0
aiomysql==0.0.20
bitstring==3.1.6
aiodns==2.0.0
uvloop==0.14.0
aioredis==1.3.0
pyyaml==5.1.2
aioredis-lock==0.0.3
python-dotenv==0.10.3
python-rapidjson==0.9.1
nanopy==20.0.0
aiohttp==3.6.2
pyyaml==5.1.2
pycryptodome==3.9.4
aiounittest==1.3.1
websockets==8.1
setuptools==42.0.2